pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "A difference calculator is a program that determines the difference between two data structures."
authors = ["BogdanBarylo <barylob@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]
classifiers = ["Programming Language :: Python :: 3.10",
               "License :: Freeware ",
               "Operating System :: OS Independent"]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^5.4.1" 

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


